Move Base class to a separate module
