+++ gaphor/UML/copypaste.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Iterator, NamedTuple
 
-from gaphor.core.modeling.element import Id
+from gaphor.core.modeling.base import Id
 from gaphor.diagram.copypaste import (
     ElementCopy,
     copy,
