+++ gaphor/core/modeling/__init__.py
@@ -1,5 +1,6 @@
 # ruff: noqa: F401,F403
 
+from gaphor.core.modeling.base import Base, Id, swap_element_type
 from gaphor.core.modeling.coremodel import (
     Comment,
     Dependency,
@@ -15,7 +16,7 @@ from gaphor.core.modeling.diagram import (
     DrawContext,
     UpdateContext,
 )
-from gaphor.core.modeling.element import Element, self_and_owners, swap_element_type
+from gaphor.core.modeling.element import Element, self_and_owners
 from gaphor.core.modeling.elementfactory import ElementFactory
 from gaphor.core.modeling.event import *
 from gaphor.core.modeling.presentation import Presentation
