+++ gaphor/diagram/presentation.py
@@ -10,8 +10,8 @@ from gaphas.constraint import constraint
 from gaphas.geometry import Rectangle, distance_rectangle_point
 from gaphas.solver.constraint import BaseConstraint
 
+from gaphor.core.modeling.base import Id
 from gaphor.core.modeling.diagram import Diagram, DrawContext
-from gaphor.core.modeling.element import Id
 from gaphor.core.modeling.event import AttributeUpdated, RevertibleEvent
 from gaphor.core.modeling.presentation import Presentation, S, literal_eval
 from gaphor.core.modeling.properties import attribute
